# specs/001-georectification/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ALPROJ GUI Backend API
  version: 1.0.0
  description: |
    景観写真オルソ化アプリのバックエンド API。
    フロントエンド（Svelte）から呼び出される。
servers:
  - url: http://localhost:{port}
    description: ローカルサイドカーサーバー
    variables:
      port:
        default: "8765"
...
paths:
...
  /api/jobs/{jobId}:
    get:
      tags: [jobs]
      summary: ジョブ状態取得
      operationId: getJobStatus
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: ジョブ状態
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

    delete:
      tags: [jobs]
      summary: ジョブキャンセル
      operationId: cancelJob
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: キャンセル成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
components:
  schemas:
...
    ProcessResult:
      type: object
      properties:
        gcps:
          type: array
          items:
            $ref: "#/components/schemas/GCP"
        metrics:
          $ref: "#/components/schemas/ProcessMetrics"
        geotiff_path:
          type: string
        log:
          type: array
          items:
            type: string
    GCP:
      type: object
      required: [id, image_x, image_y, geo_x, geo_y, geo_z, enabled]
      properties:
        id:
          type: integer
        image_x:
          type: number
        image_y:
          type: number
        geo_x:
          type: number
        geo_y:
          type: number
        geo_z:
          type: number
        residual:
          type: number
        enabled:
          type: boolean
          default: true
    ProcessMetrics:
      type: object
      required: [rmse, gcp_count, gcp_total]
      properties:
        rmse:
          type: number
        gcp_count:
          type: integer
        gcp_total:
          type: integer
        residual_mean:
          type: number
        residual_std:
          type: number
        residual_max:
          type: number

    # === Simulation ===
...
    Job:
      type: object
      required: [id, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        progress:
          type: number
          minimum: 0
          maximum: 1
        step:
          type: string
          description: 現在のステップ名
        message:
          type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        error:
          type: string
        result:
          $ref: "#/components/schemas/ProcessResult"
...